Add alternative name for interface used for kube-vip
When using heterogeneous hosts for masters the default network interface
names might differ. This causes issues with kube-vip as it uses a single
network interface name for the daemonset and if that interface doesn't
exist on the given host, then the pod will fail to start.
This can be circumvented by setting an alternative name for the
interface which kube-vip then can successfully reference.

fixes #2