Add role to create users on the NAS for access control
For users to be able to use the network shares, there actually need to
be users. Set these up with a nologin shell since there's no need for
them to actually log in and store their home folders on the ZFS pool.

Define users in the `users.yml` var file in the following format:
```yaml
users:
  user1:
    name: user1
    password: Sup4rSecret!
    groups: timemachine
```

fixes #21