Switch flux to use `flux-target` tag instead of branches
The automation introduced in d04927c automatically tags the latest commit with this tag. This means that when working on different branches flux will automatically use changes from those. To avoid introducing changes from certain branches the workflow can easily be disabled for a branch or a filter can be added to whitelist certain branches.