Add init script to create users and databases
While adding users to pgpool a3588f5 is needed for them to connect to
*pgpool*, this doesn't actually create those users in the actual
database.

To solve that adding this initdb script, that will run every time pgpool
comes up and will make sure the custom users configured for pgpool are
also created in the database along with their respective databases.