Add `service-replicated-longhorn` storageclass
This storageClass has a single replica managed by Longhorn as its purpose is to be used by services that handle replication by themselves. Other than that it is also set with `best-effort` data locality to improve performance, as this will cause Longhorn to allocate the volume on the same node as the workload itself whenever possible.