Monitor all pod/service monitors in the cluster
By default prometheus will only monitor resources that are created by
its helm chart or match the specified selector. In some charts it's not
possible to specify a label for the pod/service monitors it creates
making it impossible for them to be picked up by prometheus. By making
this change prometheus will not use the selector to pick up what to
monitor but will monitor all resources in the cluster.